renamed LearningsDao to ItemsDao
